requirements.txt
streamlit
langchain
langchain-openai
langchain-groq
mysql-connector-python
mysql
python-dotenv
